requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
python-engineio==4.9.1
python-nmap==0.7.1
python-socketio==5.11.4
requests==2.32.3
requests-oauthlib==2.0.0
simple-websocket==1.0.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
urllib3==2.2.2
URLObject==2.4.3
Werkzeug==3.0.4
wsproto==1.2.0
